requirements.txt
Flask==2.2.2
Flask-Session==0.4.0
python-dotenv==0.21.0
requests==2.28.1

